+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [FEATURE] Display mount information #73
 * [CHANGE] Implement paged node listing #133
+* [FEATURE] Undo move/clone via snackbar #42
 
 
 ## 3.1.1